refactor(profile): restructure user profile logic into modular components
- Separated user profile logic into distinct components:
  • ProfileDetail: handles user profile visualization
  • ProfileFunctions: manages edit and delete profile actions
  • GenerateToken: component for user token generation
- Improved code readability, maintainability, and component reusability.
